# GS:SW: Use doubles for Z