… pcsx2/GS/Renderers/SW/GSDrawScanlineCodeGenerator.all.cpp
… pcsx2/GS/Renderers/SW/GSDrawScanlineCodeGenerator.all.h
… pcsx2/GS/Renderers/SW/GSNewCodeGenerator.h
… pcsx2/GS/Renderers/SW/GSRasterizer.cpp
… pcsx2/GS/Renderers/SW/GSRasterizer.h
… pcsx2/GS/Renderers/SW/GSRendererSW.cpp
… pcsx2/GS/Renderers/SW/GSScanlineEnvironment.h
… pcsx2/GS/Renderers/SW/GSSetupPrimCodeGenerator.all.cpp
… pcsx2/GS/Renderers/SW/GSSetupPrimCodeGenerator.all.h
… pcsx2/GS/Renderers/SW/GSVertexSW.h