### Description of Changes
Makes the SW renderer use doubles for Z

<del>Has a noticeable (~3%) performance hit, so leaving this a draft for now, will only use it for large z values in the future</del> I think we've gotten performance to an acceptable level

### Rationale behind Changes
Reduce Z fighting

### Suggested Testing Steps
Test KOF, other zequal games / games with z fighting