@refractionpcsx2 2022-05-08
Speed losses on games which are fixed by this PR:
Destroy All Humans is fixed, and the perf loss is about 4%, but we're talking at over 210fps in both cases
Tony Hawk is about 6% but again, we're talking about 250fps here
KoF 94 fixed with about a 3.4% perf loss (but again, like 600fps down to 580fps lol)

Random other game speed losses:
GT4 147 -> 135fps (8.2%)
DT Racer 146 -> 135fps  (7.6%)
Driv3r 107 -> 101 (5.4%)
CoD 2 107 -> 100 (6.6%)

NFS MW was identical speeds (within margin of error)

Potential issues to make note of or test things from

Meta: #3692 

Individual issues:
#5780
#3496
#3609 (but awaiting feedback on some issues)
#5394
#4116