@iracusgithub 2022-05-08
This Pull Request works just fine with the majority of games I tested on.

However, theoretically, this float->to double approach might not be enough if a game abuse the floating-point arithmetic precision (kind of like what we have on the VUs or the COP1).

I have 2 questions regarding that:

Does the GS have a wider range of floats and/or follow the IEEE 754 standard for this?

And will a soft-float approach, considered in the future even if, it will likely be very,very slow?