@Lagahan 2022-05-08
Improves Quake 3
Master:

https://user-images.githubusercontent.com/5718411/167296320-58c3494c-bab0-4a4e-bc8e-198209b8f79c.mp4


PR:

https://user-images.githubusercontent.com/5718411/167296332-c9b5b4ee-f5e0-42fb-a2f5-dbe74f473188.mp4