@githubribbon45 2022-05-08
Hasbro Family Game Night is also fixed, the game table does not exhibit weird artifacts as seen in the issue report (https://github.com/PCSX2/pcsx2/issues/2289):

From the report:
![35777821-cb9ac7fc-09b4-11e8-9980-dbf912d95639](https://user-images.githubusercontent.com/105116470/167296541-54173a12-61b2-42ea-b4bd-a843a9413dab.png)

With the PR:
https://user-images.githubusercontent.com/105116470/167296571-f801cc7c-7b65-4639-8c73-02b361e63d23.mp4