@refractionpcsx2 2022-05-08
@githubribbon45

> However, theoretically, this float->to double approach might not be enough if a game abuse the floating-point arithmetic precision (kind of like what we have on the VUs or the COP1).

Not a scenario that exists (explained below)

> I have 2 questions regarding that:
> 
> Does the GS have a wider range of floats and/or follow the IEEE 754 standard for this?

The GS doesn't use floats, it's completely integer (using 12.4 FIxed integers at most), we do that for speed and convenient of interpolating coordinates.

> And will a soft-float approach, be considered in the future even if, it will likely be very,very slow?

No need here.