@JordanTheToaster 2022-05-08
Sonic Hero's no longer has Z fighting during the opening gameplay.