@ps1freak26 2022-05-08
Ridge Racer 5's shine is back in software mode so #5575 is fixed.