@prafullpcsx2 2022-05-08
This pr seems to break game "Trapt". Initial screens are not visible in SW mode. Block dump attached.

[Trapt (2022-04-09 01-09-40).zip](https://github.com/PCSX2/pcsx2/files/8647666/Trapt.2022-04-09.01-09-40.zip)