@MarioSonic2987 2022-05-10
Fixes Z-fighting in Worms 4 Mayhem. It's more noticeable in battles.

**Before**:
![image](https://user-images.githubusercontent.com/20027681/167725178-583dcfee-bfde-4faf-9aaf-85000edf29b3.png)
**After**:
![image](https://user-images.githubusercontent.com/20027681/167725189-21f8289b-455f-403d-bdbb-673660b424e4.png)