@MrCK1 2022-05-11
Confirmed that z-fighting in Shox is fixed

PR:
![image](https://user-images.githubusercontent.com/16314399/167747571-41452102-ddcd-47e4-8318-61a3b4bc9ed9.png)

Master:
![image](https://user-images.githubusercontent.com/16314399/167747411-5cccd6a3-172f-40f7-84c1-447f0cc7447e.png)
****