@TransparentBlue 2022-05-12
Does not improve Okami ( #2004 ) at all.
Does not Improve Gran Turismo 4 at all ( #3692 )
Improves Ape Escape 3 there's still some Zfigthing but you have to go out of your way to trigger it. ( #3692 )