@TransparentBlue 2022-05-12
Sorry about that, shoud've been #2008, though it's also linked in the meta issue for Zfighting.