@refractionpcsx2 2022-05-12
and to be clear, you're only testing Software mode yes? Hardware won't be affected.