@refractionpcsx2 2022-05-12
cool, thanks for the clarification :)